+++ .github/workflows/test.yml
@@ -12,8 +12,7 @@ jobs:
     name: 'Version Bump'
     runs-on: 'ubuntu-latest'
     steps:
-      - name: ${{ secrets.JENKINS_GITHUB_PAT }}
-        uses: actions/checkout@v4
+      - uses: actions/checkout@v4
         with:
           token: ${{ secrets.JENKINS_GITHUB_PAT }}
           fetch-depth: 0
@@ -28,7 +27,6 @@ jobs:
           ./package/version.sh bump ${og_version}
           ./package/version.sh sub
           new_version=$(cat package/version)
-          sed --in-place "s/^VERSION: Final = '.*'$/VERSION: Final = '${new_version}'/" kmir/src/kmir/__init__.py
           git add --update && git commit --message "Set Version: $(cat package/version)" || true
       - name: 'Push updates'
         run: git push origin HEAD:${GITHUB_HEAD_REF}
