+++ deps/k_release
@@ -1,0 +1,1 @@
+7.1.3
