+++ src/kmir/__init__.py
@@ -1,0 +1,3 @@
+from typing import Final
+
+VERSION: Final = '0.3.0'
